A good commit message example
The purpose of this commit is to provide an example of a hand-crafted,
artisanal commit message. The first line is a short, approximately 50-character
summary, followed by an empty line. The subsequent paragraphs are jam-packed
with descriptive information about the change, but each line is kept under 72
characters in length.

If even more information is needed to explain the change, more paragraphs can
be added after blank lines, with links to issues, tickets, or bugs. Remember
that future you will thank current you for your thoughtfulness and foresight!